+++ wnfs-wasm/src/fs/public/file.rs
@@ -107,7 +107,7 @@ impl PublicFile {
                 .await
                 .map_err(error("Cannot determine file size"))?;
 
-            Ok(value!(size))
+            Ok(value!(size as usize))
         }))
     }
